Decouple gesture tracking from WindowRecorder
