+++ sentry-android-replay/src/main/java/io/sentry/android/replay/WindowRecorder.kt
@@ -23,16 +23,11 @@ import kotlin.LazyThreadSafetyMode.NONE
 internal class WindowRecorder(
     private val options: SentryOptions,
     private val screenshotRecorderCallback: ScreenshotRecorderCallback? = null,
-    private val touchRecorderCallback: TouchRecorderCallback? = null,
     private val mainLooperHandler: MainLooperHandler
-) : Recorder {
+) : Recorder, OnRootViewsChangedListener {
 
     internal companion object {
         private const val TAG = "WindowRecorder"
-    }
-
-    private val rootViewsSpy by lazy(NONE) {
-        RootViewsSpy.install()
     }
 
     private val isRecording = AtomicBoolean(false)
@@ -43,15 +38,11 @@ internal class WindowRecorder(
         Executors.newSingleThreadScheduledExecutor(RecorderExecutorServiceThreadFactory())
     }
 
-    private val onRootViewsChangedListener = OnRootViewsChangedListener { root, added ->
+    override fun onRootViewsChanged(root: View, added: Boolean) {
         if (added) {
             rootViews.add(WeakReference(root))
             recorder?.bind(root)
-
-            root.startGestureTracking()
         } else {
-            root.stopGestureTracking()
-
             recorder?.unbind(root)
             rootViews.removeAll { it.get() == root }
 
@@ -68,11 +59,10 @@ internal class WindowRecorder(
         }
 
         recorder = ScreenshotRecorder(recorderConfig, options, mainLooperHandler, screenshotRecorderCallback)
-        rootViewsSpy.listeners += onRootViewsChangedListener
         capturingTask = capturer.scheduleAtFixedRateSafely(
             options,
             "$TAG.capture",
-            0L,
+            100L,
             1000L / recorderConfig.frameRate,
             MILLISECONDS
         ) {
@@ -88,7 +78,6 @@ internal class WindowRecorder(
     }
 
     override fun stop() {
-        rootViewsSpy.listeners -= onRootViewsChangedListener
         rootViews.forEach { recorder?.unbind(it.get()) }
         recorder?.close()
         rootViews.clear()
@@ -103,55 +92,6 @@ internal class WindowRecorder(
         capturer.gracefullyShutdown(options)
     }
 
-    private fun View.startGestureTracking() {
-        val window = phoneWindow
-        if (window == null) {
-            options.logger.log(DEBUG, "Window is invalid, not tracking gestures")
-            return
-        }
-
-        if (touchRecorderCallback == null) {
-            options.logger.log(DEBUG, "TouchRecorderCallback is null, not tracking gestures")
-            return
-        }
-
-        val delegate = window.callback
-        window.callback = SentryReplayGestureRecorder(options, touchRecorderCallback, delegate)
-    }
-
-    private fun View.stopGestureTracking() {
-        val window = phoneWindow
-        if (window == null) {
-            options.logger.log(DEBUG, "Window was null in stopGestureTracking")
-            return
-        }
-
-        if (window.callback is SentryReplayGestureRecorder) {
-            val delegate = (window.callback as SentryReplayGestureRecorder).delegate
-            window.callback = delegate
-        }
-    }
-
-    private class SentryReplayGestureRecorder(
-        private val options: SentryOptions,
-        private val touchRecorderCallback: TouchRecorderCallback?,
-        delegate: Window.Callback?
-    ) : FixedWindowCallback(delegate) {
-        override fun dispatchTouchEvent(event: MotionEvent?): Boolean {
-            if (event != null) {
-                val copy: MotionEvent = MotionEvent.obtainNoHistory(event)
-                try {
-                    touchRecorderCallback?.onTouchEvent(copy)
-                } catch (e: Throwable) {
-                    options.logger.log(ERROR, "Error dispatching touch event", e)
-                } finally {
-                    copy.recycle()
-                }
-            }
-            return super.dispatchTouchEvent(event)
-        }
-    }
-
     private class RecorderExecutorServiceThreadFactory : ThreadFactory {
         private var cnt = 0
         override fun newThread(r: Runnable): Thread {
@@ -160,8 +100,4 @@ internal class WindowRecorder(
             return ret
         }
     }
-}
-
-public interface TouchRecorderCallback {
-    fun onTouchEvent(event: MotionEvent)
 }
