+++ sentry-android-replay/src/main/java/io/sentry/android/replay/capture/BaseCaptureStrategy.kt
@@ -23,6 +23,7 @@ import io.sentry.android.replay.ScreenshotRecorderConfig
 import io.sentry.android.replay.capture.CaptureStrategy.Companion.createSegment
 import io.sentry.android.replay.capture.CaptureStrategy.Companion.currentEventsLock
 import io.sentry.android.replay.capture.CaptureStrategy.ReplaySegment
+import io.sentry.android.replay.gestures.ReplayGestureConverter
 import io.sentry.android.replay.util.PersistableLinkedList
 import io.sentry.android.replay.util.gracefullyShutdown
 import io.sentry.android.replay.util.submitSafely
@@ -56,15 +57,12 @@ internal abstract class BaseCaptureStrategy(
 
     internal companion object {
         private const val TAG = "CaptureStrategy"
-
-        // rrweb values
-        private const val TOUCH_MOVE_DEBOUNCE_THRESHOLD = 50
-        private const val CAPTURE_MOVE_EVENT_THRESHOLD = 500
     }
 
     private val persistingExecutor: ScheduledExecutorService by lazy {
         Executors.newSingleThreadScheduledExecutor(ReplayPersistingExecutorServiceThreadFactory())
     }
+    private val gestureConverter = ReplayGestureConverter(dateProvider)
 
     protected val isTerminating = AtomicBoolean(false)
     protected var cache: ReplayCache? = null
@@ -94,9 +92,6 @@ internal abstract class BaseCaptureStrategy(
         persistingExecutor,
         cacheProvider = { cache }
     )
-    private val currentPositions = LinkedHashMap<Int, ArrayList<Position>>(10)
-    private var touchMoveBaseline = 0L
-    private var lastCapturedMoveEvent = 0L
 
     protected val replayExecutor: ScheduledExecutorService by lazy {
         executor ?: Executors.newSingleThreadScheduledExecutor(ReplayExecutorServiceThreadFactory())
@@ -169,7 +164,7 @@ internal abstract class BaseCaptureStrategy(
     }
 
     override fun onTouchEvent(event: MotionEvent) {
-        val rrwebEvents = event.toRRWebIncrementalSnapshotEvent()
+        val rrwebEvents = gestureConverter.convert(event, recorderConfig)
         if (rrwebEvents != null) {
             synchronized(currentEventsLock) {
                 currentEvents += rrwebEvents
@@ -196,126 +191,6 @@ internal abstract class BaseCaptureStrategy(
             val ret = Thread(r, "SentryReplayPersister-" + cnt++)
             ret.setDaemon(true)
             return ret
-        }
-    }
-
-    private fun MotionEvent.toRRWebIncrementalSnapshotEvent(): List<RRWebIncrementalSnapshotEvent>? {
-        val event = this
-        return when (event.actionMasked) {
-            MotionEvent.ACTION_MOVE -> {
-                // we only throttle move events as those can be overwhelming
-                val now = dateProvider.currentTimeMillis
-                if (lastCapturedMoveEvent != 0L && lastCapturedMoveEvent + TOUCH_MOVE_DEBOUNCE_THRESHOLD > now) {
-                    return null
-                }
-                lastCapturedMoveEvent = now
-
-                currentPositions.keys.forEach { pId ->
-                    val pIndex = event.findPointerIndex(pId)
-
-                    if (pIndex == -1) {
-                        // no data for this pointer
-                        return@forEach
-                    }
-
-                    // idk why but rrweb does it like dis
-                    if (touchMoveBaseline == 0L) {
-                        touchMoveBaseline = now
-                    }
-
-                    currentPositions[pId]!! += Position().apply {
-                        x = event.getX(pIndex) * recorderConfig.scaleFactorX
-                        y = event.getY(pIndex) * recorderConfig.scaleFactorY
-                        id = 0 // html node id, but we don't have it, so hardcode to 0 to align with FE
-                        timeOffset = now - touchMoveBaseline
-                    }
-                }
-
-                val totalOffset = now - touchMoveBaseline
-                return if (totalOffset > CAPTURE_MOVE_EVENT_THRESHOLD) {
-                    val moveEvents = mutableListOf<RRWebInteractionMoveEvent>()
-                    for ((pointerId, positions) in currentPositions) {
-                        if (positions.isNotEmpty()) {
-                            moveEvents += RRWebInteractionMoveEvent().apply {
-                                this.timestamp = now
-                                this.positions = positions.map { pos ->
-                                    pos.timeOffset -= totalOffset
-                                    pos
-                                }
-                                this.pointerId = pointerId
-                            }
-                            currentPositions[pointerId]!!.clear()
-                        }
-                    }
-                    touchMoveBaseline = 0L
-                    moveEvents
-                } else {
-                    null
-                }
-            }
-
-            MotionEvent.ACTION_DOWN,
-            MotionEvent.ACTION_POINTER_DOWN -> {
-                val pId = event.getPointerId(event.actionIndex)
-                val pIndex = event.findPointerIndex(pId)
-
-                if (pIndex == -1) {
-                    // no data for this pointer
-                    return null
-                }
-
-                // new finger down - add a new pointer for tracking movement
-                currentPositions[pId] = ArrayList()
-                listOf(
-                    RRWebInteractionEvent().apply {
-                        timestamp = dateProvider.currentTimeMillis
-                        x = event.getX(pIndex) * recorderConfig.scaleFactorX
-                        y = event.getY(pIndex) * recorderConfig.scaleFactorY
-                        id = 0 // html node id, but we don't have it, so hardcode to 0 to align with FE
-                        pointerId = pId
-                        interactionType = InteractionType.TouchStart
-                    }
-                )
-            }
-            MotionEvent.ACTION_UP,
-            MotionEvent.ACTION_POINTER_UP -> {
-                val pId = event.getPointerId(event.actionIndex)
-                val pIndex = event.findPointerIndex(pId)
-
-                if (pIndex == -1) {
-                    // no data for this pointer
-                    return null
-                }
-
-                // finger lift up - remove the pointer from tracking
-                currentPositions.remove(pId)
-                listOf(
-                    RRWebInteractionEvent().apply {
-                        timestamp = dateProvider.currentTimeMillis
-                        x = event.getX(pIndex) * recorderConfig.scaleFactorX
-                        y = event.getY(pIndex) * recorderConfig.scaleFactorY
-                        id = 0 // html node id, but we don't have it, so hardcode to 0 to align with FE
-                        pointerId = pId
-                        interactionType = InteractionType.TouchEnd
-                    }
-                )
-            }
-            MotionEvent.ACTION_CANCEL -> {
-                // gesture cancelled - remove all pointers from tracking
-                currentPositions.clear()
-                listOf(
-                    RRWebInteractionEvent().apply {
-                        timestamp = dateProvider.currentTimeMillis
-                        x = event.x * recorderConfig.scaleFactorX
-                        y = event.y * recorderConfig.scaleFactorY
-                        id = 0 // html node id, but we don't have it, so hardcode to 0 to align with FE
-                        pointerId = 0 // the pointerId is not used for TouchCancel, so just set it to 0
-                        interactionType = InteractionType.TouchCancel
-                    }
-                )
-            }
-
-            else -> null
         }
     }
